interim linux compile fixes for the in64 fixes (hah)

+++ qw/source/sv_recorder.c
@@ -48,6 +48,7 @@ static __attribute__ ((used)) const char rcsid[] =
 #include <sys/time.h>
 #endif
 #include <time.h>
+#include <stdint.h>
 
 #include "QF/sizebuf.h"
 
@@ -128,7 +129,7 @@ static byte     msg_buffer[2][MAX_DATAGRAM];
 				 rec.dbuffer.start - rec.dbuffer.end : \
 				 rec.dbuffer.maxsize - rec.dbuffer.end)
 
-#define HEADER  ((size_t)(intptr_t) &((header_t *) 0)->data)
+#define HEADER  ((int) (intptr_t) &((header_t *) 0)->data)
 
 static void
 dbuffer_init (dbuffer_t *dbuffer, byte *buf, size_t size)
